Decision commit

+++ src/main/java/school/mjc/stage0/loops/task2/PowerOfTwo.java
@@ -6,7 +6,7 @@ public class PowerOfTwo {
     public void printPower(int power) {
         int counter=0;
         while(counter<=power){
-            System.out.println(Math.pow(2,counter));
+            System.out.println((int)Math.pow(2,counter));
             counter++;
         }
         if(power<0){
